bugs modal EditUser arrumado.
